admin: Move EditDialog and Selection into editDialog/ subfolder
Groups all EditDialog-related source files (EditDialog, EditDialogApiContext,
Selection, SelectionApi, SelectionRoute, Selected) and their stories and tests
into a dedicated src/editDialog/ directory.
