+++ packages/admin/admin/src/editDialog/EditDialog.tsx
@@ -19,11 +19,11 @@ import {
 } from "react";
 import { useIntl } from "react-intl";
 
-import { CancelButton } from "./common/buttons/cancel/CancelButton";
+import { CancelButton } from "../common/buttons/cancel/CancelButton";
+import { messages } from "../messages";
+import { SaveBoundary } from "../saveBoundary/SaveBoundary";
+import { SaveBoundarySaveButton } from "../saveBoundary/SaveBoundarySaveButton";
 import type { CloseDialogOptions, IEditDialogApi } from "./EditDialogApiContext";
-import { messages } from "./messages";
-import { SaveBoundary } from "./saveBoundary/SaveBoundary";
-import { SaveBoundarySaveButton } from "./saveBoundary/SaveBoundarySaveButton";
 import type { ISelectionApi } from "./SelectionApi";
 import { useSelectionRoute } from "./SelectionRoute";
