+++ packages/admin/admin/src/editDialog/__stories__/EditDialogInDataGrid.stories.tsx
@@ -5,14 +5,14 @@ import type { ReactNode } from "react";
 import { FormattedMessage } from "react-intl";
 import { useLocation } from "react-router";
 
-import { Button } from "../common/buttons/Button";
-import { FillSpace } from "../common/FillSpace";
-import { MainContent } from "../common/MainContent";
-import { DataGridToolbar } from "../common/toolbar/DataGridToolbar";
-import { useDataGridRemote } from "../dataGrid/useDataGridRemote";
+import { Button } from "../../common/buttons/Button";
+import { FillSpace } from "../../common/FillSpace";
+import { MainContent } from "../../common/MainContent";
+import { DataGridToolbar } from "../../common/toolbar/DataGridToolbar";
+import { useDataGridRemote } from "../../dataGrid/useDataGridRemote";
+import { FinalForm } from "../../FinalForm";
+import { TextField } from "../../form/fields/TextField";
 import { useEditDialog } from "../EditDialog";
-import { FinalForm } from "../FinalForm";
-import { TextField } from "../form/fields/TextField";
 
 export default {
     title: "components/edit-dialog",
