+++ packages/admin/admin/src/editDialog/__stories__/EditDialogInRouterTabs.stories.tsx
@@ -4,16 +4,16 @@ import { DataGrid, type GridToolbarProps } from "@mui/x-data-grid";
 import { type ReactNode, type RefObject, useRef } from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
-import { Button } from "../common/buttons/Button";
-import { FillSpace } from "../common/FillSpace";
-import { MainContent } from "../common/MainContent";
-import { DataGridToolbar } from "../common/toolbar/DataGridToolbar";
+import { Button } from "../../common/buttons/Button";
+import { FillSpace } from "../../common/FillSpace";
+import { MainContent } from "../../common/MainContent";
+import { DataGridToolbar } from "../../common/toolbar/DataGridToolbar";
+import { FinalForm } from "../../FinalForm";
+import { TextField } from "../../form/fields/TextField";
+import { messages } from "../../messages";
+import { RouterTab, RouterTabs } from "../../tabs/RouterTabs";
 import { EditDialog } from "../EditDialog";
 import type { IEditDialogApi } from "../EditDialogApiContext";
-import { FinalForm } from "../FinalForm";
-import { TextField } from "../form/fields/TextField";
-import { messages } from "../messages";
-import { RouterTab, RouterTabs } from "../tabs/RouterTabs";
 
 export default {
     title: "components/edit-dialog",
