+++ packages/admin/admin/src/editDialog/__stories__/MultipleEditDialogs.stories.tsx
@@ -1,9 +1,9 @@
 import { Add, Edit } from "@comet/admin-icons";
 import { Stack, Typography } from "@mui/material";
 
-import { Button } from "../common/buttons/Button";
+import { Button } from "../../common/buttons/Button";
+import { SubRoute } from "../../router/SubRoute";
 import { useEditDialog } from "../EditDialog";
-import { SubRoute } from "../router/SubRoute";
 
 const FirstComponent = () => {
     const [EditDialog, selection, editDialogApi] = useEditDialog();
